refactor: migrated processimage logic to its own helper function
